+++ backend/src/main/java/coffeeshout/room/application/service/RoomService.java
@@ -13,6 +13,7 @@ import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.event.MiniGameSelectEvent;
 import coffeeshout.room.domain.event.PlayerListUpdateEvent;
 import coffeeshout.room.domain.event.PlayerReadyEvent;
+import coffeeshout.room.domain.event.QrCodeStatusEvent;
 import coffeeshout.room.domain.event.RoomCreateEvent;
 import coffeeshout.room.domain.event.RoomJoinEvent;
 import coffeeshout.room.domain.player.Player;
@@ -221,6 +222,36 @@ public class RoomService {
         );
 
         eventPublisher.publishEvent(new PlayerListUpdateEvent(event.joinCode()));
+    }
+
+    public void handleQrCodeStatus(QrCodeStatusEvent event) {
+        log.info(
+                "QR 코드 완료 이벤트 수신: eventId={}, joinCode={}, status={}",
+                event.eventId(), event.joinCode(), event.status()
+        );
+
+        switch (event.status()) {
+            case SUCCESS -> {
+                log.info(
+                        "QR 코드 완료 이벤트 처리 완료 (SUCCESS): eventId={}, joinCode={}, url={}",
+                        event.eventId(), event.joinCode(), event.qrCodeUrl()
+                );
+                roomCommandService.assignQrCode(new JoinCode(event.joinCode()), event.qrCodeUrl());
+                eventPublisher.publishEvent(event);
+            }
+            case ERROR -> {
+                log.info(
+                        "QR 코드 완료 이벤트 처리 완료 (ERROR): eventId={}, joinCode={}",
+                        event.eventId(), event.joinCode()
+                );
+                roomCommandService.assignQrCodeError(new JoinCode(event.joinCode()));
+                eventPublisher.publishEvent(event);
+            }
+            default -> log.error(
+                    "처리할 수 없는 QR 코드 상태: eventId={}, joinCode={}, status={}",
+                    event.eventId(), event.joinCode(), event.status()
+            );
+        }
     }
 
     private void saveRoomEntity(String joinCodeValue) {
